+++ inst/tinytest/test-ragg.R
@@ -17,23 +17,26 @@ expect_snapshot_plot(p1, "ragg-base")
 
 # bad plot always fails
 if (Sys.info()["sysname"] %in% getOption("tinysnapshot_os", default = Sys.info()["sysname"])) {
-    expect_false(ignore(expect_snapshot_plot)(p2, "ragg-base"))
+    expect_false(ignore(expect_snapshot_plot)(p2, "ragg-base", review = FALSE))
 }
 
 ###### ggplot2
 suppressPackageStartupMessages(library("ggplot2"))
 
 if (!ON_CRAN) {
-    p1 <- ggplot(mtcars, aes(mpg, hp)) + geom_point()
+    p1 <- ggplot(mtcars, aes(mpg, hp)) +
+        geom_point()
     expect_snapshot_plot(p1, "ragg-ggplot2_variable")
 }
 
-p2 <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
+p2 <- ggplot(mtcars, aes(mpg, wt)) +
+    geom_point()
 if (Sys.info()["sysname"] %in% getOption("tinysnapshot_os", default = Sys.info()["sysname"])) {
-    expect_false(ignore(expect_snapshot_plot)(p2, "ragg-ggplot2_variable"))
+    expect_false(ignore(expect_snapshot_plot)(p2, "ragg-ggplot2_variable", review = FALSE))
 }
 
 
 
 
 options(tinysnapshot_device = NULL)
+